backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pydantic
aiofiles
python-dotenv
httpx
typing-extensions 